requirements.txt
setuptools
pywinauto
pyautogui
pygetwindow
customtkinter
openai
Pillow  # This is the PIL (Python Imaging Library) package
speechRecognition
requests
pywin32
psutil
ImageGrab  # Part of Pillow
fuzzywuzzy
pytesseract
uiautomation
gtts
pygame
pyAudio
